Explicit use of database transactions
We know where we want to use transactions or not: functions and methods
are designed to modify database or just query it.

Reflect that knowledge in functions and method signatures instead of
trying to hide it behind an interface.

It could maybe be valuable to define an interface for read-only DB
access that actually enforces the RO property (here we use *sql.DB but
code can call Exec() on it). It would only define the Query/QueryRow
methods.