Upgrade the db schema upgrade process
Version the schema, and look for update scripts in a directory. Update
scripts are name "00001.sql", "00002.sql", etc. Each number correspond
to a version. No database means version 0. Each script is expected to
add a record in the version table.

On database open, check that our current version (in a table) is up to
date, and execute the scripts for each missing version in order.